fix: clean up DataProcess, Data
